Fix fn selector builder for custom types
Before, the function selector used for methods with
custom types was being built as `"my_func(MyStruct)"`.

This isn't the right approach, instead, it should be:
`"my_func(s(field_1,field_2,...,field_n))"`.

This fixes the function selector builder to support the
right approach and it also greatly simplifies the whole
process.