Add support for contract calls with struct as inputs
Before the changes introduced here, using the SDK to call contracts
that takes structs as input would cause weird bugs -- it didn't really
cause a panic on the VM, but it made the access to fields within
a struct retrieve garbage data.

That was happening because in the `script_data` we sent in the tx
we weren't including a `call_data_offset` that's computed as:

```rust
let call_data_offset = script_data_offset as usize
                        + ContractId::LEN + 2 * WORD_SIZE;
```

Once that was included, struct field access worked perfectly.

Curiously, this `call_data_offset` isn't needed if the input is a
primitive type (u8, bool, ...). So a conditional check that adds
the `call_data_offset` to the `script_data` iff the ABI method called
takes a struct is now in place.

This enables users to successfully perform the following:

```Rust
// ...
let contract_instance = MyContract::new(compiled, client);

let counter_config = CounterConfig {
    dummy: true,
    initial_value: 42,
};

let result = contract_instance
    .initialize_counter(counter_config)
    .call()
    .await
    .unwrap();

assert_eq!(42, result);
```