Handle return data larger than a word
When the returned data from a contract call is larger than a word,
it isn't stored in `Return.val`, but in `ReturnData.data`.

Before the changes in here, we were pulling the data only from
`Return.val`, which fortunately handled most cases. Until now,
with some methods returning, for instance, a `b256`, which would cause
the data to _not_ be in `Return.val` as we expected, leading to errors.

This PR introduces some changes to fix this by looking whether the
expected return type is larger than a word, if it is, we fetch the
returned data from `ReturnData.data` instead of `Return.val`.