Add `fuels-types` minimal crate for declaring serializable ABI types.
By moving these from `sway-types` upstream to the `fuels-rs` repo, we
can resolve the remainder of the cyclic dependency between the `sway`
and `fuels-rs` repositories.

I think semantically it also makes more sense for these VM-specific
types to be declared upstream, and for Sway to depend on these rather
than vice versa.