Specify minimum supported Rust version in `fuels-rs` manifest
This ensures that users are warned in the case that their installed Rust
version predates the version required by `fuels-rs`.

This should be particularly useful in `forc`-generated harnesses, as I
believe we've already had a couple of reports of strange compile errors
that just required a Rust version update.

You can find documentation on this field [here][1]:

[1]: https://doc.rust-lang.org/cargo/reference/manifest.html#the-rust-version-field