Distinguish between a `Wallet`'s `Locked` and `Unlocked` states
This PR was motivated by work @kayagokalp and I are doing on the
following:
https://github.com/FuelLabs/sway/issues/2413

We noticed that in order to perform any operations using a `Wallet`, the
private key must be known, even if those operations were simply
retrieving asset balances or a list of transactions.

This PR aims to draw a clearer distinction between Wallet methods that
require access to a private key and those that do not. The idea is to
encourage users of the API to be conscious of when their wallet is in an
unlocked state (i.e. their private key resides in memory) and when it
does not.

To achieve this, we turn `Wallet` into a session type, parameterised
over its state which can be either `Locked` or `Unlocked`.

The default state is `Locked` in the case that the parameter is not
specified for `Wallet` or `LocalWallet`, meaning that if users require
holding the wallet in an unlocked state they must specify the type
explicitly, e.g. `Wallet<Unlocked>`. I think this explicitness can
assist us in maintaining awareness of the state of the wallet that we
require in our downstream APIs.

Users should generally prefer using the `Wallet` in a locked state (the
default) and should ideally only `unlock` the wallet for the duration
that they need to perform operations requiring the private key. Users
can now construct a `Wallet` in a `Locked` state, e.g.
`Wallet::from_address(public_addr, provider)` and access the various
useful read-only operations without knowing the private key or having it
stored in memory.

As a follow-up to this work, we should take a closer look at the usage
of `LocalWallet` in downstream crates (including `fuels-contract`) to
see if we can refactor these to use the wallet in a `Locked` state more
often, only moving into the `Unlocked` state when absolutely necessary.