Simplify wallet access by removing the state type input parameter
This removes the input state type from `Wallet<T>` in favour of using
two simple types, `Wallet` and `WalletUnlocked`.

`WalletUnlocked` provides a `Deref` implementation that targets its
inner `Wallet`, allowing it to be used with API's that take a `&Wallet`
as an input and to call the `Wallet` methods directly.

The `LocalWallet` has been removed as a part of this commit in order to
unnecessarily introduce a new `LocalWalletUnlocked` type when we may
wish to remove the ambiguous `LocalWallet` type in either way: see #543.