Fuel Core v0.16.1 (#813)
Update `fuels-rs` to the latest `fuel-core 0.16.1` and latest `sway
0.34.0`.

Due to changes in `fuel-asm` for GTF, this is causing yet another
circular upgrade issue.
- Used `{ workspace = true }` syntax to use one version across the
project and simplify integration with new changes in the future.
- Added `ci_checks.sh` script to compile all tests to simplify
integration of changes like this in the future.
- Update contracts to be compatible with `sway 0.34.0`.
- Updated tests to be compatible with `fuel-core 0.16.1`.

related PR: https://github.com/FuelLabs/sway/pull/3961

---------

Co-authored-by: green <xgreenx9999@gmail.com>