feat!: `beta-5` support (#1200)
BREAKING CHANGE:
- `TxParameters` are replaced with `TxPolicies`
- `GasPrice` and `Maturity` fields are optional 
- `TxPolicies` introduced new fields: 
- `WitnessLimit` - allows the limitation of the maximum size of
witnesses in bytes.
- `MaxFee` - allows the upper bound for the maximum fee that users agree
to pay for the transaction.
- The `ScriptGasLimit` only limits the script execution. Previously, the
`ScriptGasLimit` also limited the predicate execution time, but it is
not valid anymore. So, it is not possible to use the `ScriptGasLimit`
for transaction cost limitations. A new `MaxFee` policy is a way to do
that. The `GasLimit` field was removed from the `Create` transaction
because it only relates to the script execution(which the `Create`
transaction doesn't have).
- The new `WhitnessLimit` also impacts the `max_gas` and `max_fee`
calculation along with the `ScriptGasLimit`(in the case of `Create`
transaction only `WitnessLimit` affects the `max_gas` and `max_fee`).
- The minimal gas also charges the user for transaction ID calculation.
- Each transaction requires setting the `GasPrice` policy.
- Previously, `GasLimit` should be less than the `MAX_GAS_PER_TX`
constant. After removing this field from the `Create` transaction, it is
impossible to require it. Instead, it requires that `max_gas <=
MAX_GAS_PER_TX` for any transaction. Consequently, any `Script`
transaction that uses `MAX_GAS_PER_TX` as a `GasLimit` will always fail
because of a new rule. Setting the estimated gas usage instead solves
the problem.
- If the `max_fee > policies.max_fee`, then transaction will be
rejected.
- If the `witnessses_size > policies.witness_limit`, then transaction
will be rejected.
- `get_message_proof` not uses `Nonce` instead of the message_id
- predicates do not use `ChainId` for address calculation
- `manual_blocks_enabled` is replaced with `debug` in local chain config
- `fee_checked_from_tx` uses `FeeParameters`
- `fuel_tx::ConsensusParameters` were refactored - this also affects us
- when building a transaction_builder the `BuildableTransacion` trait
needs to be in scope
- `utxo_validation` and `manual_blocks` are enabled by default for test
providers
- node config does not have `local_node` anymore use `default` - `let
node_config = Config::default();`

Thanks @MujkicA for updating the documentation.

Co-authored-by: MujkicA <ahmedmujkic2@gmail.com>
Co-authored-by: Ahmed Sagdati <37515857+segfault-magnet@users.noreply.github.com>