refactor!: use `submit_and_await_commit` API (#1187)
Closes #1180,  #1128.
This PR uses the correct API for sending a transaction and ensuring that
we wait for the commit of the transaction before returning.
This is needed because otherwise there might be a disconnect between
submitting the transaction and awaiting for its commit, if it is done
using the two separate APIs as before.

BREAKING CHANGE: 
- `send_transaction_and_await_commit` function now returns a `TxStatus`
instead of `TxId`.

---------

Co-authored-by: MujkicA <32431923+MujkicA@users.noreply.github.com>
Co-authored-by: hal3e <git@hal3e.io>